Simplify implementation of Sign and Handedness
This lets Sign/Handedness be instantiated directly as Sign(-1),
Handedness(1) etc., which is potentially unsafe, but ensures that a
constructor/prototype are created in JavaScript (to allow instanceof to
work as expected) and allows for destructuring in Scala. In some cases
it may also be more efficient for client code to construct a
Sign/Handedness object from an integer directly.