Refactor direction types
Switch back to basing on components, not a vector; this is easier now
that directions are more distinct from vectors, and should be a bit more
efficient in most cases. Also get rid of 'None' constants and disallow
creating 'null' directions from a zero vector.

In the process, simplify Rotation3d a bit to be more consistent with
Rotation2d (hide basis computation inside class).