Start normalizing interval operations
Interval operations will be updated to follow Double operations more
closely. For any given operation, the pattern will be:
- determine all possible Double values resulting from applying that
  operation to operands (or pairs of operands) within the given
  interval(s)
- strip out NaN values
- if the resulting set is empty, the result is Interval.Empty; otherwise
  the result is Interval(min(results), max(results))

For instance, for 'Double.PositiveInfinity - Interval.Whole':
- the only possible result values are NaN (Double.PositiveInfinity -
  Double.PositiveInfinity) and Double.PositiveInfinity
  (Double.PositiveInfinity - <anything else>)
- after stripping out NaN, the only possible result is
  Double.PositiveInfinity so the result is
  Interval(Double.PositiveInfinity, Double.PositiveInfinity)

Interval(0,0) will be considered to be Interval(-0,+0). Interval(0, x) for
x > 0 will be considered to be Interval(+0, x). Interval(x, 0) for x < 0
will be considered to be Interval(-x, -0). (This is important when
considering division.)

This may lead to slightly surprising results in some cases but is
consistent with Double arithmetic, easy to explain/apply and relatively
efficient (since in many cases simply using Double arithmetic on interval
endpoints will 'do the right thing').